Expose libvirt/hw_machine_type for consistent machine types.
The libvirt driver currently does not set the machine type for a KVM
guest by default.  When not specified, libvirt will use the newest
one it knows about.  Unfortunately, that can result in live
migrations failing if your environment is using different versions of
the host OS on compute nodes as the destination node may not be able
to support the machine type used when the VM was originally started.

Change-Id: I269ade8997389fc56ce3009fbd0660d1a47a0c4a