Set cpu_mode to "none" when virt_type is qemu
The Nova code states that the default value for cpu_mode when using
qemu as virt_type should be host-model [1]. However, this is not
actually tested in the gate due to [2], and with QEMU 2.6.0 it is
known to cause issues [3], because there is a chance a new CPU
flag is used, but not emulated by QEMU TCG code.

Setting cpu_mode to "none" in this case as default would prevent us
from hitting issues when this code is most exercised, i.e. CI
environments.

[1] https://github.com/openstack/nova/blob/f419825c216bccd9f7023deec024099c5b65428d/nova/conf/libvirt.py#L442-L448
[2] https://github.com/openstack-dev/devstack/blob/master/lib/nova_plugins/hypervisor-libvirt#L42
[3] https://bugzilla.redhat.com/show_bug.cgi?id=1404627

Change-Id: Ic95755a927eef0580433cd5c8eca1c4c7cad2f69