Allow db sync timeouts to be configurable
As Openstack projects continue to have longer database migration
chains, the Puppet default timeout of 300 seconds for an execution
is becoming too short a duration on some hardware, leading to timeouts.
As projects continue to add more migration scripts without "pruning"
the base, timeouts will continue to become more frequent unless
this time can be expanded.

The proposed change adds a configuration option db_sync_timeout
which maintains the Puppet default of 300 but allows TripleO to
change the timeout to something higher.

Change-Id: I6b30a4d9e3ca25d9a473e4eb614a8769fa4567e7