Make placement's os_interface configurable
this exposes the os_interface option for the placement API
configuration, which enables us to set the interface (public, internal
or admin) to use for the placement API endpoint. Before, it was
hardcoded to public, due to the keystoneauth1 library's defaults.
The change was introduced by Ic996e596f8473c0b8626e8d0e92e1bf58044b4f8

Change-Id: I1c7fd3a32d04e2fafb3820d1c1f221f45c613c83
Related-Bug: #1663187