Use newer configuration options for libvirt's live migration URI
The live_migration_uri option that was being used in case TLS is used in
live migration doesn't permit the usage of a host other than the one
detected automatically by nova. This is problematic for cases like TLS
where the address needs to match the certificate's SubjectAltName entry.
So instead, this patch takes into use the two recommended options:

* live_migration_inbound_addr
* live_migration_scheme

These will be used by nova to determine the URI.

On the other hand, libvirt_migration_uri is set to be deprecated at some
point. But, a way to add the query parameters to the query is
needed before this happens.

Change-Id: I0a1684397ebefaa8dc00237e0b7952e9296381fa