Properly handle arrays for enabled_perf_events
Since the nova_config provider assumes that arrays are multi-str
configuration items, the enabled_perf_events would not be properly
configured if an array was provided. This resulted in idempotency issues
when [] was used as well as dropping values if ['cmt','mbml'] was
passed. This change ensures a comma seperated list is used for the
configuration value of enabled_perf_events.

Change-Id: I41ad1d4a7f14acae10e673676cd8adc98eddbec5
Closes-Bug: #1690946
(cherry picked from commit 8ed5c5e974b35c65bdee617e9881c833179a7b7c)