Add pci/alias to compute manifest
Starting with Ocata pci_alias needs to be configured on compute nodes as
well.

This requires refactoring of the manifests as the parameter is common to
nova::compute and nova::api.
Common pci configuration is now done by the nova::pci class while compute
specific configuration is done now by the nova::compute::pci class.

Closes-Bug: #1702730
Change-Id: Iac4f22cbe043629ae9f0477ae62cf77ad0d761f8