Allow creating security group rules for ICMP
At the moment, it's not possible to create a security group
rule with from port and to port set to -1.  This is useful
only when creating ICMP rules to allow all ICMP traffic.

This patch allows setting both values to -1, only if the
protocol of the security group rule is ICMP.

Change-Id: I290005b31fd4afc246db28ffd899302fb85a67fb
(cherry picked from commit dad40312ebc12c3f79ef71fc0649342c102b59fa)