Switch nova_security_rule to openstack provider
The current nova provider for nova_security_rule is broken at the
moment.  Due to the fact that the commands are getting deprecated,
the warnings are confusing the text parsing and result in repeated
creation of security group rules.

This patch resolves this issue by switching it to the new openstack
provider.  It also adds the instances and prefetch methods which
will allow `puppet resource` CLI usage.

Change-Id: I9f84728f89a8344685f64b79c8079f3bf74ff979