Change Nova Conductor workers to ::os_workers
In Ieb69b5a6a50d37b0a49a8c9e57275673d60b5f10, the default
number of workers for the metadata_workers and osapi_compute_workers
was changed from $::processorcount to $::os_workers, which takes
CPU count into account but caps the number of workers for
machines with high numbers of CPUs.   The
"workers" configuration option used by Nova Conductor however
was left out of this, likely because it did not previously
default to $::processorcount.  However Nova itself defaults
all of these services to the processor count in any case,
and the number of workers for Conductor is likely the most critical
value to be limited, since it directly corresponds to pooled
database connections which when there are too many processes
results in often thousands of connections that are almost all
unused.   This patch limits Conductor workers in the same
way as that of the other worker counts.

Change-Id: I1969b555c1c8cfb686a12cad7f54ac10b1656af5