Add group configuration to qemu.conf file
Vhost user sockets are shared between qemu and ovs. Currently,
we are patching ovs service file to make ovs run as qemu user.
In order to avoid it, a new group 'hugetlbfs' has been introduced,
which will be used by ovs and qemu. This patch allows the users
to configure the group setting for qemu.conf file.

Change-Id: I0fa3934b345d51de85fd9d8a5a211b8e1bc50d86