Support --purge in Nova cleanup crons.
This patch adds the --purge parameter from
'nova-manage db' when executing the cleanup
tasks.

It will add the --purge parameter to the
archive_deleted_rows cron and will add an additional
cron to exclusively purge the shadow tables.

This feature is relevant due to the fact that we
were not purging the databases before running
upgrades, leading to the increase of the DB sizes.

Change-Id: I4ac0c9bd30463c546b141326b1293c78f845cf81