Add a retention policy (age) when purging shadow tables
This submission changes the --all default parameter in the purge shadow tables
cron to a '--before <date>' allowing to define a retention policy. Age name was
used as puppet-cinder already use it so to have a consitancy across the
projects.

Also, it's added the option --all-cells causing the purge to be applied
against all cell databases.

Co-Authored-By: Sergii Golovatiuk <sgolovat@redhat.com>

Change-Id: I2dcf37417c36fb8b1bde207c60d22d580005715c