Set valid_interfaces for placement
The os_interface config option in the placement
section was removed and has no effect so if
anybody is using the os_interface parameter
for this class it doesn't do anything.

See release notes [1]. This changes so that the
puppet interface is intact but the correct config
option is set.

This is fixed for Stein with [2].

[1] https://docs.openstack.org/releasenotes/nova/rocky.html
[2] https://review.openstack.org/#/c/605069/

Change-Id: I869af3ebd37691e0117e320661bf9edabce4efd6