Set DEFAULT/enabled_apis in nova::api
enabled_apis [1] won't get set in nova::api when
nova_metadata_wsgi_enabled == true.
But the enabled_apis list get also modified for this case, so
we need to move setting the config outside the conditional at [3].
Since the config is right now also set in nova::metadata this
deprecates it in nova::metadata to only use nova::api.

Related-Bug: #1832537

[1] https://github.com/openstack/puppet-nova/blob/master/manifests/api.pp#L401
[2] https://github.com/openstack/puppet-nova/blob/master/manifests/api.pp#L353
[3] https://github.com/openstack/puppet-nova/blob/master/manifests/api.pp#L379

Change-Id: Ifd751feecb4bdbece138fec3a7c90ba5377bd0d0