Don't restart libvirt-guests on config change
When there is a config change the service gets notfied to be
restarted, which is not required as /usr/libexec/libvirt-guests.sh
sources /etc/sysconfig/libvirt-guests on each run. As a side
effect of the service restart, instances get stopped which is bad.

In addition manage_service was bound to the enabled parameter.
This introduces manage_service to be able to manage start/stop
independent from enabled parameter.

Change-Id: Icf52a2272601401c423ae07f7b10474bdc190574
Closes-Bug: #1849264