Migrate back cross_az_attach
The previous change[1] migrated some parameters related cinder
from nova to nova::cinder, but cross_az_attach shouldn't be migrate
because the parameter is used in nova-api and nova-compute and in some
cases nova::cinder class is not actually loaded in api side.
(We actually don't include nova::cinder for controllers in TripleO)

This patch partially reverts the change by [1], so that we keep
cross_az_attach in more common place, which is surely loaded in
both control plane and hypervisor.

[1] 72103db985d00b3289b8b936956166f20ef8f3d0

Change-Id: I2e07e29740cf0398b4f34457b53d54890b4aa843