Fix dependency cycle
There is a dependency cycle when using the nova::compute::libvirt
class and the nova::compute::rbd class in the same manifest.

The 'set-secret-value virsh' resource says before the
nova::config::end resource but up it's dependency path the
new libvirtd-tcp sources introduced in [1] depends requires
the nova::config::end anchor.

This removes the before nova::config::end on the set-secret-value
resources and only makes it depend on the get-or-set resource.

2020-06-05 08:26:58.966645 | centos-8 | (Anchor[nova::config::end] =>
Service[libvirtd-tcp] => Service[libvirt] => Exec[get-or-set virsh
secret] => Exec[set-secret-value virsh] =>
Anchor[nova::config::end])\nTry the '--graph' option and opening the
resulting '.dot' file in OmniGraffle or GraphViz

[1] https://review.opendev.org/#/c/730689/

Change-Id: Iab96968de75d7064b3aecf266d0f24b7b98dc809