Support libvirt/cpu_models instead of libvirt/cpu_model
The libvirt/cpu_model parameter has been deprecated in nova[1] in
favor of the new parameter, cpu_models, so puppet-nova should support
that new parameter instead of the old one.

Note that this change directly removes cpu_model parameter which is
added by the parent change, to avoid a redundant deprecated parameter
created in a single release.

[1] f80e5f989db7f326e824b671cfe0a8bbb047cfca
Change-Id: Ib8e9ec2e0bde59e72b0ce3639b64d3e8a6ff9eac