Migrate scheduler-generic parameters to nova::scheduler
This patch migrates some parameters from nova::scheduler::filter to
nova::scheduler, because these are parameters to determine behavior of
nova-scheduler, and not specific to scheduler filters.

Note that the default value for max_attempts parameter has been changed
from 3 to $::os_service_default, because its default is defined as 3 in
nova implementation.

Change-Id: Ic74e1fdb4adf9f954b6e58343050c5e166d40889