Fix libvirt version discovery for CentOS Stream
CentOS Stream is Rolling-release distro that tracks just ahead of
Red Hat Enterprise Linux (RHEL) development [1]. While not recommended
in RDO, it's used to find potential issues before they are found in
official CentOS 8.

CentOS Stream is reported by facter as '8' with no minor version,
breaking the logic to discover the libvirtd version. This patch is
setting the version of libvirtd to 5.6 for all CdentOS 8 releases,
including Stream, as 8.0 is no longer supported.

[1] https://www.centos.org/centos-stream/

Change-Id: I468ecf06ac6e0853fbb296c011f20d63e38f6922