Fix incomplete backport of e8b24805d6ad863d91ecd84d00333e14821a7368
The change 7301812c9e32cde7ee145a1cbd18e8e76f52be37 was supposed to
be a backport of e8b24805d6ad863d91ecd84d00333e14821a7368 but in fact
it was incomplete and didn't have required changes in
the nova::metadata class.

This change implements the missing part to backport the change
completely.

Change-Id: Ie9620ddecc1ded6659b75be6acfd7de125c9b8ba