Override default mkisofs_cmd in CentOS 9.
In [1] the parameter mkisofs_cmd was introduced in puppet-nova to set
this parameter as CentOS Stream 9 requires a different value that the
default one.

This change overrides the default value in CentOS9 so that no
additional tuning is needed.

[1] https://review.opendev.org/c/openstack/puppet-nova/+/810932

Change-Id: Id35db695a350fccad74688c96555393d0c108a8b
(cherry picked from commit 39d206911997c4208baae6f2220e7756fdb985c7)