Support [mdev_<type>] mdev_class
Since Xena release, nova supports defining device class for each mdev
type[1]. This change introduces the new mdev_types parameter and
the resource type support for the parameter.

[1] 9be996c696e91e82d11d81db72ffb4cc151598b6

The existing interface is still kept, to make this change backportable
to older stable branches, but will be deprecated in master later.

Change-Id: I7a7d22eb9aa27cebabb2cf04a7f1451741c6d4c3