Wallaby-only: Fix handling of [DEFAULT] max_concurrent_builds
The change bf1e6fd9c187af56911beb7797588b731efe7a04 introduced
the logic to handle the deprecated max_concurrent_builds parameter but
the logic only works when nova::compute::ironic is included BEFORE
nova::compute. This is not always true and can break the existing
usage.

This change ensures the deprecated parameter works regardless of
the order to include classes. One disadvantage of this change is that
the parameter is not longer cleared out by default, but users can still
ensure it is removed from nova.conf by setting $::os_service_default
explicitly.

The deprecated parameter was removed during stable/xena cycle so this
is submitted as a wallaby-only change.

Change-Id: Ib2d8eaaef0105f32de8efe78b2a7b73a4f1f4201