vmware: Fix the wrong package installed
All common implementation for vmware has been split out to
the oslo.vmware library, which is currently an optional dependency.
We have to make sure this library is installed so that vmware driver
is functional.

Also, python(3)-suds is required by python(3)-oslo-vmware so we don't
have to explicitly installs this library.

Additionally this fixes the wrong indent in the spec file.

Closes-Bug: #2011828
Change-Id: Ica93dd5127e6330e0d5a65738ca68c43e8be3058