Deprecate client class
The nova CLI has been deprecated[1] and will be removed in a future
release. This deprecates the class to install the novaclient package
because this class aims to help installing the package for CLI usage.
It should be installed by package dependencies in case it is used as
a library.

[1] 0fb7190c062a08a3969a71955c014e3a572cdc9e

Change-Id: Iffe2cc8daf28d088c098443986925c18a8412318