libvirt: Accept more flexible types for [libvirt] hw_machine_type
The [libvirt] hw_machine_type option accepts a string value which
represents a hash value in the foromat below.

hw_machine_type=key1=value1,key2=value2

This introduces the logic to accept a list value or a hash value so
that users can defined the expected value in a more "native" format.

Change-Id: I6eb060d0621d688cd95820378323a5d80cc08ab4