Avoid deprecated live_migration_uri if possible
The live_migration_uri option was deprecated in favor of the new
live_migration_(scheme|inbound_addr) options.

Note that the new options does not allow customizing some parameters
for ssh, which were possible by the deprecated _uri option, so
the legacy option is still used if a user requests any of the options
currently not configurable by the new options.

Related-Bug: #1680729
Change-Id: I88c9777999cb091b8570c642e75b0332fb103294