Use secret_ceph resource in nova::compute::rbd
This migrates the libvirt secret handling in the
nova::compute::rbd class to using the new
secret_ceph definition.

We also deprecate the rbd_keyring approach were
we now require the usage of libvirt_rbd_secret_key
if manage_libvirt_secret is set to true (the default)
and we have a libvirt_rbd_secret_uuid given.

Change-Id: I8be56ebfc7c7ed6fb6d1d92f0ffec4a6d714a5d1